# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo List Api",
        "description": "Demo Todo List Api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/labtechnicians": {
            "get": {
                "tags": [
                    "Lab Technicians"
                ],
                "summary": "Get all lab technicians",
                "description": "Retrieve all lab technicians.",
                "operationId": "6bc58f26ef1cd5820729153a3962174d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "labtechnicians": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LabTechnician"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Lab Technicians"
                ],
                "summary": "Store a new lab technician",
                "description": "Store a new lab technician.",
                "operationId": "4e95a7f22b2629de3a5565fdafa91572",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "speciality",
                                    "image"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "speciality": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/labtechnicians/by-speciality": {
            "get": {
                "tags": [
                    "Lab Technicians"
                ],
                "summary": "Get lab technicians by speciality",
                "description": "Get lab technicians by speciality.",
                "operationId": "e1607257297a39ace073cec8880771eb",
                "parameters": [
                    {
                        "name": "speciality",
                        "in": "query",
                        "description": "Speciality of lab technicians",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "labtechnicians": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LabTechnician"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/labtechnicians/{id}": {
            "get": {
                "tags": [
                    "Lab Technicians"
                ],
                "summary": "Get a specific lab technician",
                "operationId": "176b73c7cfe338bef0825716baac955d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the lab technician",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "labtechnician": {
                                            "$ref": "#/components/schemas/LabTechnician"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Lab technician not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Lab Technicians"
                ],
                "summary": "Update a lab technician",
                "description": "Update a lab technician.",
                "operationId": "a041a93c21a7cb0e5cb8d3ed5bdd43d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LabTechnician"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lab Technicians"
                ],
                "summary": "Delete a lab technician",
                "description": "Delete a lab technician.",
                "operationId": "6e3d5bf8d8b02dd31dba44259defbf29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LabTechnician": {
                "required": [
                    "name",
                    "phone",
                    "speciality",
                    "image"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "phone": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "speciality": {
                        "type": "string",
                        "example": "Laboratory Technician"
                    },
                    "image": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}